requirements_backend.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
PyJWT
pytz
python-dotenv
django-rest-knox
